+++ example/lib/sources/signup_form.dart
@@ -54,8 +54,9 @@ class _SignupFormState extends State<SignupForm> {
                   decoration: InputDecoration(
                     labelText: 'Confirm Password',
                     suffixIcon: (_formKey.currentState != null &&
-                            !_formKey.currentState.fields['confirm_password']
-                                .isValid)
+                            !(_formKey.currentState?.fields['confirm_password']
+                                    ?.isValid ??
+                                false))
                         ? const Icon(Icons.error, color: Colors.red)
                         : const Icon(Icons.check, color: Colors.green),
                   ),
@@ -68,7 +69,7 @@ class _SignupFormState extends State<SignupForm> {
                             : null),*/
                     (val) {
                       if (val !=
-                          _formKey.currentState.fields['password'].value) {
+                          _formKey.currentState?.fields['password']?.value) {
                         return 'Passwords do not match';
                       }
                       return null;
@@ -84,7 +85,7 @@ class _SignupFormState extends State<SignupForm> {
                   ]),
                   // initialValue: true,
                   decoration: InputDecoration(labelText: 'Accept Terms?'),
-                  builder: (FormFieldState<bool> field) {
+                  builder: (FormFieldState<bool?> field) {
                     return InputDecorator(
                       decoration: InputDecoration(
                         errorText: field.errorText,
@@ -104,17 +105,14 @@ class _SignupFormState extends State<SignupForm> {
                 MaterialButton(
                   color: Theme.of(context).accentColor,
                   onPressed: () {
-                    if (_formKey.currentState.saveAndValidate()) {
+                    if (_formKey.currentState?.saveAndValidate() ?? false) {
                       print('Valid');
                     } else {
                       print('Invalid');
                     }
-                    print(_formKey.currentState.value);
+                    print(_formKey.currentState?.value);
                   },
-                  child: Text(
-                    'Signup',
-                    style: TextStyle(color: Colors.white),
-                  ),
+                  child: Text('Signup', style: TextStyle(color: Colors.white)),
                 )
               ],
             ),