+++ lib/src/fields/form_builder_date_range_picker.dart
@@ -11,71 +11,72 @@ import 'package:intl/intl.dart' as intl;
 class FormBuilderDateRangePicker extends FormBuilderField<DateTimeRange> {
   //TODO: Add documentation
   final int maxLines;
-  final TextInputType keyboardType;
+  final TextInputType? keyboardType;
   final bool obscureText;
-  final TextStyle style;
-  final TextEditingController controller;
+  final TextStyle? style;
+  final TextEditingController? controller;
   final TextCapitalization textCapitalization;
-  final TextInputAction textInputAction;
-  final StrutStyle strutStyle;
-  final TextDirection textDirection;
+  final TextInputAction? textInputAction;
+  final StrutStyle? strutStyle;
+  final TextDirection? textDirection;
   final TextAlign textAlign;
   final bool autofocus;
   final bool autocorrect;
-  final MaxLengthEnforcement maxLengthEnforcement;
-  final int maxLength;
-  final VoidCallback onEditingComplete;
-  final ValueChanged<List<DateTime>> onFieldSubmitted;
-  final List<TextInputFormatter> inputFormatters;
+  final MaxLengthEnforcement? maxLengthEnforcement;
+  final int? maxLength;
+  final VoidCallback? onEditingComplete;
+  final ValueChanged<List<DateTime>>? onFieldSubmitted;
+  final List<TextInputFormatter>? inputFormatters;
   final double cursorWidth;
-  final Radius cursorRadius;
-  final Color cursorColor;
-  final Brightness keyboardAppearance;
+  final Radius? cursorRadius;
+  final Color? cursorColor;
+  final Brightness? keyboardAppearance;
   final EdgeInsets scrollPadding;
   final bool enableInteractiveSelection;
-  final InputCounterWidgetBuilder buildCounter;
+  final InputCounterWidgetBuilder? buildCounter;
   final bool expands;
-  final int minLines;
+  final int? minLines;
   final bool showCursor;
   final DateTime firstDate;
   final DateTime lastDate;
-  final Locale locale;
-  final intl.DateFormat format;
-  final String cancelText; // widget.cancelText,
-  final String confirmText; // widget.confirmText,
-  final DateTime currentDate; // widget.currentDate,
-  final String errorFormatText; // widget.erroerrorFormatText,
-  final Widget Function(BuildContext, Widget) pickerBuilder; // widget.builder,
-  final String errorInvalidRangeText; // widget.errorInvalidRangeText,
-  final String errorInvalidText; // widget.errorInvalidText,
-  final String fieldEndHintText; // widget.fieldEndHintText,
-  final String fieldEndLabelText; // widget.fieldEndLabelText,
-  final String fieldStartHintText; // widget.fieldStartHintText,
-  final String fieldStartLabelText; // widget.fieldStartLabelText,
-  final String helpText; // widget.helpText,
+  final Locale? locale;
+  final intl.DateFormat? format;
+  final String? cancelText; // widget.cancelText,
+  final String? confirmText; // widget.confirmText,
+  final DateTime? currentDate; // widget.currentDate,
+  final String? errorFormatText; // widget.erroerrorFormatText,
+  final Widget Function(BuildContext, Widget?)?
+      pickerBuilder; // widget.builder,
+  final String? errorInvalidRangeText; // widget.errorInvalidRangeText,
+  final String? errorInvalidText; // widget.errorInvalidText,
+  final String? fieldEndHintText; // widget.fieldEndHintText,
+  final String? fieldEndLabelText; // widget.fieldEndLabelText,
+  final String? fieldStartHintText; // widget.fieldStartHintText,
+  final String? fieldStartLabelText; // widget.fieldStartLabelText,
+  final String? helpText; // widget.helpText,
   // final DateTimeRange initialDateRange; // widget.initialDateRange,
   final DatePickerEntryMode initialEntryMode; // widget.initialEntryMode,
-  final RouteSettings routeSettings; // widget.routeSettings,
-  final String saveText; // widget.saveText,
+  final RouteSettings? routeSettings; // widget.routeSettings,
+  final String? saveText; // widget.saveText,
   final bool useRootNavigator; // widget.useRootNavigator,
 
   /// Creates field for selecting a range of dates
   FormBuilderDateRangePicker({
-    Key key,
+    Key? key,
     //From Super
-    @required String name,
-    FormFieldValidator<DateTimeRange> validator,
-    DateTimeRange initialValue,
+    required String name,
+    FormFieldValidator<DateTimeRange>? validator,
+    DateTimeRange? initialValue,
     InputDecoration decoration = const InputDecoration(),
-    ValueChanged<DateTimeRange> onChanged,
-    ValueTransformer<DateTimeRange> valueTransformer,
+    ValueChanged<DateTimeRange>? onChanged,
+    ValueTransformer<DateTimeRange>? valueTransformer,
     bool enabled = true,
-    FormFieldSetter<DateTimeRange> onSaved,
+    FormFieldSetter<DateTimeRange>? onSaved,
     AutovalidateMode autovalidateMode = AutovalidateMode.disabled,
-    VoidCallback onReset,
-    FocusNode focusNode,
-    @required this.firstDate,
-    @required this.lastDate,
+    VoidCallback? onReset,
+    FocusNode? focusNode,
+    required this.firstDate,
+    required this.lastDate,
     this.format,
     this.maxLines = 1,
     this.obscureText = false,
@@ -103,7 +104,7 @@ class FormBuilderDateRangePicker extends FormBuilderField<DateTimeRange> {
     this.buildCounter,
     this.expands = false,
     this.minLines,
-    this.showCursor,
+    this.showCursor = false,
     this.locale,
     this.cancelText,
     this.confirmText,
@@ -135,7 +136,7 @@ class FormBuilderDateRangePicker extends FormBuilderField<DateTimeRange> {
           onReset: onReset,
           decoration: decoration,
           focusNode: focusNode,
-          builder: (FormFieldState<DateTimeRange> field) {
+          builder: (FormFieldState<DateTimeRange?> field) {
             final state = field as FormBuilderDateRangePickerState;
 
             return TextField(
@@ -179,12 +180,10 @@ class FormBuilderDateRangePicker extends FormBuilderField<DateTimeRange> {
       FormBuilderDateRangePickerState();
 
   static String tryFormat(DateTime date, intl.DateFormat format) {
-    if (date != null) {
-      try {
-        return format.format(date);
-      } catch (e) {
-        // print('Error formatting date: $e');
-      }
+    try {
+      return format.format(date);
+    } catch (e) {
+      // print('Error formatting date: $e');
     }
     return '';
   }
@@ -192,19 +191,19 @@ class FormBuilderDateRangePicker extends FormBuilderField<DateTimeRange> {
 
 class FormBuilderDateRangePickerState
     extends FormBuilderFieldState<FormBuilderDateRangePicker, DateTimeRange> {
-  TextEditingController _effectiveController;
+  late TextEditingController _effectiveController;
 
   @override
   void initState() {
     super.initState();
     _effectiveController =
         widget.controller ?? TextEditingController(text: _valueToText());
-    effectiveFocusNode.addListener(_handleFocus);
+    effectiveFocusNode!.addListener(_handleFocus);
   }
 
   @override
   void dispose() {
-    effectiveFocusNode.removeListener(_handleFocus);
+    effectiveFocusNode!.removeListener(_handleFocus);
     // Dispose the _effectiveController when initState created it
     if (null == widget.controller) {
       _effectiveController.dispose();
@@ -213,8 +212,8 @@ class FormBuilderDateRangePickerState
   }
 
   Future<void> _handleFocus() async {
-    if (effectiveFocusNode.hasFocus && enabled) {
-      effectiveFocusNode.unfocus();
+    if (effectiveFocusNode!.hasFocus && enabled) {
+      effectiveFocusNode!.unfocus();
       /*final initialFirstDate = value?.isEmpty ?? true
           ? (widget.initialFirstDate ?? DateTime.now())
           : value[0];
@@ -256,20 +255,18 @@ class FormBuilderDateRangePickerState
       return '';
     }
 
-    return '${format(value.start)} - ${format(value.end)}';
+    return '${format(value!.start)} - ${format(value!.end)}';
   }
 
   String format(DateTime date) => FormBuilderDateRangePicker.tryFormat(
       date, widget.format ?? intl.DateFormat.yMd());
 
   void _setTextFieldString() {
-    setState(() {
-      _effectiveController.text = _valueToText();
-    });
+    setState(() => _effectiveController.text = _valueToText());
   }
 
   @override
-  void didChange(DateTimeRange value) {
+  void didChange(DateTimeRange? value) {
     super.didChange(value);
     _setTextFieldString();
   }