+++ test/fields/form_builder_tester.dart
@@ -14,5 +14,5 @@ Widget buildTestableFieldWidget(Widget widget) {
   );
 }
 
-bool formSave() => _formKey.currentState.saveAndValidate();
-dynamic formValue(String name) => _formKey.currentState.value[name];
+bool formSave() => _formKey.currentState!.saveAndValidate();
+dynamic formValue(String name) => _formKey.currentState!.value[name];